[fix] User entity 충돌 해결
