+++ src/main/java/com/example/outsourcingproject/domain/menu/repository/MenuRepository.java
@@ -1,0 +1,19 @@
+package com.example.outsourcingproject.domain.menu.repository;
+
+import com.example.outsourcingproject.domain.menu.entity.Menu;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Repository;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.util.Optional;
+
+@Repository
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+
+    Optional<Menu> findById(Long id);
+
+    default Menu findByIdOrElseThrow(Long id){
+        return findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exists id: " + id));
+    }
+}
