+++ src/main/java/com/example/outsourcingproject/domain/order/dto/response/OrderAcceptResponseDto.java
@@ -1,0 +1,13 @@
+package com.example.outsourcingproject.domain.order.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class OrderAcceptResponseDto {
+
+    private final String message;
+
+    public OrderAcceptResponseDto(String message) {
+        this.message = message;
+    }
+}
