+++ src/main/java/com/example/outsourcingproject/domain/order/dto/response/OrderRejectResponseDto.java
@@ -1,0 +1,13 @@
+package com.example.outsourcingproject.domain.order.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class OrderRejectResponseDto {
+
+    private final String message;
+
+    public OrderRejectResponseDto(String message) {
+        this.message = message;
+    }
+}
