+++ src/main/java/com/example/outsourcingproject/domain/orderItem/repository/OrderItemRepository.java
@@ -1,0 +1,14 @@
+package com.example.outsourcingproject.domain.orderItem.repository;
+
+import com.example.outsourcingproject.domain.orderItem.entity.OrderItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+
+    List<OrderItem> findAllByOrderId(Long orderId);
+
+}
