+++ src/main/java/com/example/outsourcingproject/domain/store/repository/StoreRepository.java
@@ -1,0 +1,19 @@
+package com.example.outsourcingproject.domain.store.repository;
+
+import com.example.outsourcingproject.domain.store.entity.Store;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Repository;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.util.Optional;
+
+@Repository
+public interface StoreRepository extends JpaRepository<Store, Long> {
+
+    Optional<Store> findById(Long id);
+
+    default Store findByIdOrElseThrow(Long id){
+        return findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exists id: " + id));
+    }
+}
