+++ src/main/java/com/example/outsourcingproject/domain/user/repository/UserRepository.java
@@ -1,9 +1,9 @@
 package com.example.outsourcingproject.domain.user.repository;
 
 import com.example.outsourcingproject.domain.user.entity.User;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
@@ -13,4 +13,8 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 
     Optional<User> findById(Long id);
+
+    default User findByIdOrElseThrow(Long id){
+        return findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exists id: " + id));
+    }
 }
