--- src/main/java/com/example/outsourcingproject/order/dto/request/OrderStatusRequestDto.java
@@ -1,10 +1,0 @@
-package com.example.outsourcingproject.order.dto.request;
-
-import com.example.outsourcingproject.order.enums.Status;
-import lombok.Getter;
-
-@Getter
-public class OrderStatusRequestDto {
-
-    private Status status;
-}
