--- src/main/java/com/example/outsourcingproject/order/dto/response/OrderAcceptResponseDto.java
@@ -1,13 +1,0 @@
-package com.example.outsourcingproject.order.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class OrderAcceptResponseDto {
-
-    private final String message;
-
-    public OrderAcceptResponseDto(String message) {
-        this.message = message;
-    }
-}
