--- src/main/java/com/example/outsourcingproject/order/dto/response/OrderRejectResponseDto.java
@@ -1,13 +1,0 @@
-package com.example.outsourcingproject.order.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class OrderRejectResponseDto {
-
-    private final String message;
-
-    public OrderRejectResponseDto(String message) {
-        this.message = message;
-    }
-}
