--- src/main/java/com/example/outsourcingproject/order/dto/response/OrderSaveResponseDto.java
@@ -1,13 +1,0 @@
-package com.example.outsourcingproject.order.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class OrderSaveResponseDto {
-
-    private final String message;
-
-    public OrderSaveResponseDto(String message) {
-        this.message = message;
-    }
-}
