--- src/main/java/com/example/outsourcingproject/order/dto/response/OrderStatusResponseDto.java
@@ -1,13 +1,0 @@
-package com.example.outsourcingproject.order.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class OrderStatusResponseDto {
-
-    private final String message;
-
-    public OrderStatusResponseDto(String message) {
-        this.message = message;
-    }
-}
