--- src/main/java/com/example/outsourcingproject/order/repository/OrderRepository.java
@@ -1,20 +1,0 @@
-package com.example.outsourcingproject.order.repository;
-
-import com.example.outsourcingproject.order.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Repository;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
-
-    List<Order> findAllByUserId(Long id);
-
-    default Order findByIdOrElseThrow(Long id){
-        return findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exists id: " + id));
-    }
-}
