--- src/main/java/com/example/outsourcingproject/order/validation/BasicController.java
@@ -1,20 +1,0 @@
-package com.example.outsourcingproject.order.validation;
-
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
-import org.springframework.web.bind.WebDataBinder;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.InitBinder;
-
-@ControllerAdvice
-public class BasicController {
-    protected final LocalValidatorFactoryBean validator;
-
-    public BasicController(LocalValidatorFactoryBean validator) {
-        this.validator = validator;
-    }
-
-    @InitBinder
-    public void initBinder(WebDataBinder binder) {
-        binder.addValidators(new CollectionValidator(validator));
-    }
-}
