--- src/main/java/com/example/outsourcingproject/user/repository/UserRepository.java
@@ -1,19 +1,0 @@
-package com.example.outsourcingproject.user.repository;
-
-import com.example.outsourcingproject.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Repository;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findById(Long id);
-
-    default User findByIdOrElseThrow(Long id){
-        return findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exists id: " + id));
-    }
-}
