+++ src/main/resources/application.yml
@@ -1,4 +1,4 @@
 spring:
   config:
     import:
-      - classpath:/prod/application-prod.yml
+      - classpath:/local/application-local.yml
